Better tracking.  Only 'allow' brightness tracking for small movement.
The idea is that the background subtractor will (should) only fail when
the fish is barely moving or motionless, so the fallback of using the
brightness tracker should only consider and return bright objects near
the fish's last-estimated position.  This should not change over time,
as the background subtractor should pick up on any movement that could
get the fish somewher else.